requirements.txt
asgiref==3.7.2
Django==3.2.8
djangorestframework==3.12.0
djangorestframework-api-key==2.3.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
psycopg2==2.9.1